Date filter 2

+++ topik/tests/test_models.py
@@ -1,42 +1,60 @@
 import os
 import unittest
+from abc import ABCMeta, abstractmethod
 
-from topik.readers import iter_document_json_stream
-from topik.tokenizers import SimpleTokenizer
-from topik.vectorizers import CorpusBOW
-from topik.models import LDA
-from topik.utils import unzip
+from six import with_metaclass
+
+from topik.readers import read_input
+from topik.preprocessing import preprocess
+from topik.models import LDA, PLSA
 
 # sample data files are located in the same folder
 module_path = os.path.dirname(__file__)
 
-class TestModels(unittest.TestCase):
+NTOPICS = 3
+
+
+class ModelBase(with_metaclass(ABCMeta)):
 
     def setUp(self):
-        id_documents =  iter_document_json_stream(
-                os.path.join(module_path,'data/test-data-1'), "text")
-        ids, doc_text = unzip(id_documents)
-        my_corpus = SimpleTokenizer(doc_text)
-        corpus_bow = CorpusBOW(my_corpus)
-        corpus_bow.save_dict(os.path.join(module_path, 'test.dict'))
-        corpus_bow.serialize(os.path.join(module_path, 'test.mm'))
+        raw_data = read_input(
+                source=os.path.join(module_path, 'data/test_data_json_stream.json'),
+                content_field="abstract")
+        self.digested_data = preprocess(raw_data)
+        self.model = self._train_model()
 
     def tearDown(self):
-        os.remove(os.path.join(module_path, 'test.dict'))
-        os.remove(os.path.join(module_path, 'test.mm'))
-        os.remove(os.path.join(module_path, 'test.model'))
+        if os.path.exists(os.path.join(module_path, 'test.model')):
+            os.remove(os.path.join(module_path, 'test.model'))
 
-    def test_lda(self):
-        my_lda = LDA("test.mm", "test.dict", ntopics=3)
-        my_lda.save(os.path.join(module_path, 'test.model'))
-        top_words = my_lda.get_top_words(15)
+    @abstractmethod
+    def _train_model(self):
+        raise NotImplementedError
+
+    def test_save_data(self):
+        self.model.save(os.path.join(module_path, 'test.model'))
+        self.assertTrue(os.path.isfile(os.path.join(module_path, 'test.model')))
+
+    def test_top_words(self):
+        top_words = self.model.get_top_words(15)
         self.assertEqual(len(top_words), 3)
         self.assertEqual(len(top_words[1]), 15)
 
-        my_lda.termite_data(os.path.isfile(os.path.join(module_path, 'test_termite.csv')))
+    def test_termite_output(self):
+        self.model.termite_data(os.path.join(module_path, 'test_termite.csv'))
+        self.assertTrue(os.path.isfile(os.path.join(module_path, 'test_termite.csv')))
 
-        self.assertTrue(os.path.isfile(os.path.join(module_path, 'test.model')))
-        self.assertTrue(os.path.isfile(os.path.join(module_path, 'test_termite.model')))
+
+class TestLDA(ModelBase, unittest.TestCase):
+    def _train_model(self):
+        return LDA(self.digested_data, ntopics=NTOPICS)
+
+"""
+class TestPLSA(ModelBase, unittest.TestCase):
+    def _train_model(self):
+        return PLSA(self.digested_data, topics=NTOPICS)
+"""
+
 
 if __name__ == '__main__':
     unittest.main()
